mcp-sec-10ks/pyproject.toml
[project]
name = "mcp-sec-10ks"
version = "0.1.0"
description = "MCP Server and Client for SEC 10-K Financial Data Extraction with LLM Sampling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "mcp",
    "openai",
    "edgartools",
    "pandas",
    "aiohttp",
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["mcp_sec_10ks", "client", "batch_client"]
